抽出addTask方法
