+++ src/main/java/taskflow/work/builder/WorkBuilder.java
@@ -1,8 +1,12 @@
 package taskflow.work.builder;
 
-import java.util.function.Supplier;
+import java.util.HashMap;
+import java.util.LinkedHashMap;
+import java.util.Map;
+import java.util.function.Function;
 
 import taskflow.constants.PropertyNameAndValue;
+import taskflow.task.Task;
 import taskflow.work.SequentialRouteWork;
 
 /**
@@ -15,11 +19,29 @@ public abstract class WorkBuilder {
 		//初始化默认参数
 		PropertyNameAndValue.init();
 	}
+	protected Map<String, Task> taskMap = null;
+	protected Map<String, String> taskRefExtraMap = null;
+	
+	WorkBuilder() {
+		taskMap = new LinkedHashMap<>();
+	}
+	
+	public WorkBuilder addTask(Task task) {
+		taskMap.put(task.toString(), task);
+		return this;
+	}
+	public WorkBuilder addTask(Task task,String extra) {
+		taskRefExtraMap = taskRefExtraMap == null ? new HashMap<>() : taskRefExtraMap;
+		taskRefExtraMap.put(task.toString(), extra);
+		return addTask(task);
+	}
+	
+	private static Function<Map<String, String>,SequentialRouteWork> DEFAULT_WORK_CREATER=(extraMap) -> new SequentialRouteWork(extraMap);
 	public static SequentialWorkBuilder newInstance() {
-		return new SequentialWorkBuilder(new SequentialRouteWork(null));
+		return new SequentialWorkBuilder(DEFAULT_WORK_CREATER);
 	}
 
-	public static SequentialWorkBuilder newInstance(Supplier<SequentialRouteWork> workSupplier) {
-		return new SequentialWorkBuilder(workSupplier.get());
+	public static SequentialWorkBuilder newInstance(Function<Map<String, String>,SequentialRouteWork> workCreater) {
+		return new SequentialWorkBuilder(workCreater);
 	}
 }
